message_broker/requirements.txt
kafka-python>=2.0.2,<2.1.0